Split index.js into multiple modules. Update readme section on Usage
